+++ cosmovisor/cmd/cosmovisor/cmd/help_test.go
@@ -88,95 +88,7 @@ func (s *HelpTestSuite) setEnv(t *testing.T, env *cosmovisorHelpEnv) {
 	}
 }
 
-func (s *HelpTestSuite) TestShouldGiveHelpEnvVars() {
-	initialEnv := s.clearEnv()
-	defer s.setEnv(nil, initialEnv)
-
-	emptyVal := ""
-	homeVal := "/somehome"
-	nameVal := "somename"
-
-	tests := []struct {
-		name     string
-		envHome  *string
-		envName  *string
-		expected bool
-	}{
-		{
-			name:     "home set name set",
-			envHome:  &homeVal,
-			envName:  &nameVal,
-			expected: false,
-		},
-		{
-			name:     "home not set name not set",
-			envHome:  nil,
-			envName:  nil,
-			expected: true,
-		},
-		{
-			name:     "home empty name not set",
-			envHome:  &emptyVal,
-			envName:  nil,
-			expected: true,
-		},
-		{
-			name:     "home set name not set",
-			envHome:  &homeVal,
-			envName:  nil,
-			expected: true,
-		},
-		{
-			name:     "home not set name empty",
-			envHome:  nil,
-			envName:  &emptyVal,
-			expected: true,
-		},
-		{
-			name:     "home empty name empty",
-			envHome:  &emptyVal,
-			envName:  &emptyVal,
-			expected: true,
-		},
-		{
-			name:     "home set name empty",
-			envHome:  &homeVal,
-			envName:  &emptyVal,
-			expected: true,
-		},
-		{
-			name:     "home not set name set",
-			envHome:  nil,
-			envName:  &nameVal,
-			expected: true,
-		},
-		{
-			name:     "home empty name set",
-			envHome:  &emptyVal,
-			envName:  &nameVal,
-			expected: true,
-		},
-	}
-
-	prepEnv := func(t *testing.T, envVar string, envVal *string) {
-		if envVal == nil {
-			require.NoError(t, os.Unsetenv(cosmovisor.EnvHome))
-		} else {
-			require.NoError(t, os.Setenv(envVar, *envVal))
-		}
-	}
-
-	for _, tc := range tests {
-		s.T().Run(tc.name, func(t *testing.T) {
-			prepEnv(t, cosmovisor.EnvHome, tc.envHome)
-			prepEnv(t, cosmovisor.EnvName, tc.envName)
-			actual := ShouldGiveHelp(nil)
-			assert.Equal(t, tc.expected, actual)
-		})
-	}
-}
-
-func (s HelpTestSuite) TestShouldGiveHelpArgs() {
+func (s HelpTestSuite) TestShouldGiveHelpArg() {
 	initialEnv := s.clearEnv()
 	defer s.setEnv(nil, initialEnv)
 
@@ -184,79 +96,59 @@ func (s HelpTestSuite) TestShouldGiveHelpArgs() {
 
 	tests := []struct {
 		name     string
-		args     []string
+		arg      string
 		expected bool
 	}{
 		{
-			name:     "nil args",
-			args:     nil,
+			name:     "empty string",
+			arg:      "",
 			expected: false,
 		},
 		{
-			name:     "empty args",
-			args:     []string{},
+			name:     "random",
+			arg:      "random",
 			expected: false,
 		},
 		{
-			name:     "one arg random",
-			args:     []string{"random"},
+			name:     "help",
+			arg:      "help",
+			expected: true,
+		},
+		{
+			name:     "-h",
+			arg:      "-h",
+			expected: true,
+		},
+		{
+			name:     "--help",
+			arg:      "--help",
+			expected: true,
+		},
+		{
+			name:     "help weird casing",
+			arg:      "hELP",
+			expected: true,
+		},
+		{
+			name:     "version",
+			arg:      "version",
 			expected: false,
 		},
 		{
-			name:     "five args random",
-			args:     []string{"random1", "--random2", "-r", "random4", "-random5"},
+			name:     "--version",
+			arg:      "--version",
 			expected: false,
 		},
 		{
-			name:     "one arg help",
-			args:     []string{"help"},
-			expected: true,
-		},
-		{
-			name:     " two args help first",
-			args:     []string{"help", "arg2"},
-			expected: true,
-		},
-		{
-			name:     "two args help second",
-			args:     []string{"arg1", "help"},
-			expected: true,
-		},
-		{
-			name:     "one arg -h",
-			args:     []string{"-h"},
-			expected: true,
-		},
-		{
-			name:     "two args -h first",
-			args:     []string{"-h", "arg2"},
-			expected: true,
-		},
-		{
-			name:     "two args -h second",
-			args:     []string{"arg1", "-h"},
-			expected: true,
-		},
-		{
-			name:     "one arg --help",
-			args:     []string{"--help"},
-			expected: true,
-		},
-		{
-			name:     "two args --help first",
-			args:     []string{"--help", "arg2"},
-			expected: true,
-		},
-		{
-			name:     "two args --help second",
-			args:     []string{"arg1", "--help"},
-			expected: true,
+			name:     "run",
+			arg:      "run",
+			expected: false,
 		},
 	}
 
 	for _, tc := range tests {
-		s.T().Run(tc.name, func(t *testing.T) {
-			actual := ShouldGiveHelp(tc.args)
+		s.T().Run(fmt.Sprintf("%s - %t", tc.name, tc.expected), func(t *testing.T) {
+			actual := ShouldGiveHelp(tc.arg)
 			assert.Equal(t, tc.expected, actual)
 		})
 	}