+++ crypto/keys/multisig/amino.go
@@ -64,7 +64,7 @@ func tmToProto(tmPk tmMultisig) (*LegacyAminoPubKey, error) {
 }
 
 // MarshalAminoJSON overrides amino JSON unmarshaling.
-func (m LegacyAminoPubKey) MarshalAminoJSON() (tmMultisig, error) { //nolint:golint
+func (m LegacyAminoPubKey) MarshalAminoJSON() (tmMultisig, error) { //nolint:golint,revive
 	return protoToTm(&m)
 }
 
@@ -78,11 +78,28 @@ func (m *LegacyAminoPubKey) UnmarshalAminoJSON(tmPk tmMultisig) error {
 	// Instead of just doing `*m = *protoPk`, we prefer to modify in-place the
 	// existing Anys inside `m` (instead of allocating new Anys), as so not to
 	// break the `.compat` fields in the existing Anys.
-	m.PubKeys = make([]*types.Any, len(protoPk.PubKeys))
+	if m.PubKeys == nil {
+		m.PubKeys = make([]*types.Any, len(tmPk.PubKeys))
+	}
 	for i := range m.PubKeys {
-		m.PubKeys[i] = &types.Any{}
-		m.PubKeys[i].TypeUrl = protoPk.PubKeys[i].TypeUrl
-		m.PubKeys[i].Value = protoPk.PubKeys[i].Value
+		if m.PubKeys[i] == nil {
+			// create the compat jsonBz value
+			bz, err := AminoCdc.MarshalJSON(tmPk.PubKeys[i])
+			if err != nil {
+				return err
+			}
+
+			m.PubKeys[i] = protoPk.PubKeys[i]
+			// UnmarshalJSON():
+			// just sets the compat.jsonBz value.
+			// always succeeds: err == nil
+			if err := m.PubKeys[i].UnmarshalJSON(bz); err != nil {
+				return err
+			}
+		} else {
+			m.PubKeys[i].TypeUrl = protoPk.PubKeys[i].TypeUrl
+			m.PubKeys[i].Value = protoPk.PubKeys[i].Value
+		}
 	}
 	m.Threshold = protoPk.Threshold