+++ x/auth/middleware/testutil_test.go
@@ -26,8 +26,9 @@ import (
 
 // testAccount represents an account used in the tests in x/auth/middleware.
 type testAccount struct {
-	acc  authtypes.AccountI
-	priv cryptotypes.PrivKey
+	acc    authtypes.AccountI
+	priv   cryptotypes.PrivKey
+	accNum uint64
 }
 
 // MWTestSuite is a test suite to be used with middleware tests.
@@ -42,7 +43,7 @@ type MWTestSuite struct {
 // returns context and app with params set on account keeper
 func createTestApp(t *testing.T, isCheckTx bool) (*simapp.SimApp, sdk.Context) {
 	app := simapp.Setup(t, isCheckTx)
-	ctx := app.BaseApp.NewContext(isCheckTx, tmproto.Header{}).WithBlockGasMeter(sdk.NewInfiniteGasMeter())
+	ctx := app.BaseApp.NewContext(isCheckTx, tmproto.Header{Height: app.LastBlockHeight() + 1}).WithBlockGasMeter(sdk.NewInfiniteGasMeter())
 	app.AccountKeeper.SetParams(ctx, authtypes.DefaultParams())
 
 	return app, ctx
@@ -52,7 +53,6 @@ func createTestApp(t *testing.T, isCheckTx bool) (*simapp.SimApp, sdk.Context) {
 func (s *MWTestSuite) SetupTest(isCheckTx bool) sdk.Context {
 	var ctx sdk.Context
 	s.app, ctx = createTestApp(s.T(), isCheckTx)
-	ctx = ctx.WithBlockHeight(1)
 
 	// Set up TxConfig.
 	encodingConfig := simapp.MakeTestEncodingConfig()
@@ -89,25 +89,23 @@ func (s *MWTestSuite) SetupTest(isCheckTx bool) sdk.Context {
 
 // createTestAccounts creates `numAccs` accounts, and return all relevant
 // information about them including their private keys.
-func (s *MWTestSuite) createTestAccounts(ctx sdk.Context, numAccs int) []testAccount {
+func (s *MWTestSuite) createTestAccounts(ctx sdk.Context, numAccs int, coins sdk.Coins) []testAccount {
 	var accounts []testAccount
 
 	for i := 0; i < numAccs; i++ {
 		priv, _, addr := testdata.KeyTestPubAddr()
 		acc := s.app.AccountKeeper.NewAccountWithAddress(ctx, addr)
-		err := acc.SetAccountNumber(uint64(i))
+		accNum := uint64(i)
+		err := acc.SetAccountNumber(accNum)
 		s.Require().NoError(err)
 		s.app.AccountKeeper.SetAccount(ctx, acc)
-		someCoins := sdk.Coins{
-			sdk.NewInt64Coin("atom", 10000000),
-		}
-		err = s.app.BankKeeper.MintCoins(ctx, minttypes.ModuleName, someCoins)
+		err = s.app.BankKeeper.MintCoins(ctx, minttypes.ModuleName, coins)
 		s.Require().NoError(err)
 
-		err = s.app.BankKeeper.SendCoinsFromModuleToAccount(ctx, minttypes.ModuleName, addr, someCoins)
+		err = s.app.BankKeeper.SendCoinsFromModuleToAccount(ctx, minttypes.ModuleName, addr, coins)
 		s.Require().NoError(err)
 
-		accounts = append(accounts, testAccount{acc, priv})
+		accounts = append(accounts, testAccount{acc, priv, accNum})
 	}
 
 	return accounts
@@ -139,9 +137,11 @@ func (s *MWTestSuite) createTestTx(txBuilder client.TxBuilder, privs []cryptotyp
 	sigsV2 = []signing.SignatureV2{}
 	for i, priv := range privs {
 		signerData := xauthsigning.SignerData{
+			Address:       sdk.AccAddress(priv.PubKey().Address()).String(),
 			ChainID:       chainID,
 			AccountNumber: accNums[i],
 			Sequence:      accSeqs[i],
+			SignerIndex:   i,
 		}
 		sigV2, err := tx.SignWithPrivKey(
 			s.clientCtx.TxConfig.SignModeHandler().DefaultMode(), signerData,