+++ cosmovisor/CHANGELOG.md
@@ -36,6 +36,18 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Features
+
++ [\#10285](https://github.com/cosmos/cosmos-sdk/pull/10316) Added `run` action.
+
+### Deprecated
+
++ [\#10285](https://github.com/cosmos/cosmos-sdk/pull/10316) Running `cosmovisor` without the `run` argument.
+
+### Bug Fixes
+
++ [\#10458](https://github.com/cosmos/cosmos-sdk/pull/10458) Fix version when using 'go install github.com/cosmos/cosmos-sdk/cosmovisor/cmd/cosmovisor@v1.0.0' to install cosmovisor.
+
 ## v1.0.0 2021-09-30
 
 ### Features